narrow down the directive provider interface
