+++ common/scala/src/main/resources/reference.conf
@@ -9,7 +9,7 @@ whisk.spi {
   LogStoreProvider = whisk.core.containerpool.logging.DockerToActivationLogStoreProvider
   LoadBalancerProvider = whisk.core.loadBalancer.ShardingContainerPoolBalancer
   EntitlementSpiProvider = whisk.core.entitlement.LocalEntitlementProvider
-  AuthenticationDirectiveProvider = whisk.core.controller.BasicAuthenticationDirectiveProvider
+  AuthenticationDirectiveProvider = whisk.core.controller.BasicAuthenticationDirective
 }
 
 dispatchers {
